MaidenThoughts/webapp/migrations/0002_auto_20151018_1055.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import migrations, models
import tinymce.models


class Migration(migrations.Migration):

    dependencies = [
        ('webapp', '0001_initial'),
    ]

    operations = [
        migrations.AlterField(
            model_name='articles',
            name='content',
            field=tinymce.models.HTMLField(max_length=10000000000L),
        ),
    ]
