requirements.txt
awsebcli==3.5.4
botocore==1.2.11
cement==2.4.0
colorama==0.3.3
Django==1.9.1
django-ckeditor==5.0.2
django-filebrowser==3.6.1
django-grappelli==2.7.1
django-tinymce==2.0.5
django-wysiwyg==0.7.1
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
jmespath==0.9.0
pathspec==0.3.3
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
six==1.10.0
texttable==0.8.3
websocket-client==0.32.0
wheel==0.24.0
